refactor: move velocity conversion to midi

+++ src/envelope.h
@@ -9,7 +9,7 @@ void  init_envelopes();
 void  set_envelope_attack(uint16_t attack_ms);
 void  set_envelope_decay(uint16_t decay_ms);
 void  set_envelope_release(uint16_t release_ms);
-void  trigger_envelope(lane_id id, uint8_t velocity);
+void  trigger_envelope(lane_id id, float velocity);
 void  release_envelope(lane_id id);
 float envelope_nextval(lane_id id);
 bool  envelope_is_running(lane_id id);
